add conditional before delete, to delete remaining game_characters from said table

+++ web/assignments/project1/rdiclosegame.php
@@ -29,10 +29,12 @@
         $dbUpdate = $db->prepare($updateQuery);
         $dbUpdate->execute(array(':player_count' => $result[0]['player_count'], ':game_id' => $game_id, ':host_user' => $user_id));
 
-        // delete any remaining characters from the game_characters table
-        $deleteQuery = "DELETE FROM rdi_game_characters WHERE game_id=:game_id";
-        $dbDelete = $db->prepare($deleteQuery);
-        $dbDelete->execute(array(":game_id" => $game_id));
+        if ($result[0]['joined_players'] < $result[0]['player_count']) {
+            // delete any remaining characters from the game_characters table
+            $deleteQuery = "DELETE FROM rdi_game_characters WHERE game_id=:game_id";
+            $dbDelete = $db->prepare($deleteQuery);
+            $dbDelete->execute(array(":game_id" => $game_id));
+        }
     }  else {
         // clean the output buffer
         ob_clean();
